CSS updates for H3-H6 and Long Sample Code Blocks
This update changes H3-H6 such that they are not All Caps.  It also
takes long code blocks and makes sure that it does not wrap
horizontally and has a maximum height of 50cm.  If it exceeds the view
port, then it provides vertical and horizontal scrolling.  The
documentation was also updated to reflect this.