Update instructions for AC::Metal
With 2.1.26, it is automaticly instrumented. Remove the note in
troubleshooting, and let existing users know it's a noop now, and won't
hurt anything.

The white-space=nowrap is because that snippet of code hits the end of
the width, so make it wrap as a single word, and not hyphenate.
Hyphenated code is weird.