chore: deploy automatically when merging to master
Since the tags and the version releases are controlled by the upstream
version of the repository, there is no need for versioning locally. It
might even conflict with other tags, so we would not be able to update
the fork. Once the repository is fully mature, we could start our own
clone and then version it accordingly, but for now it's not worth it.

closes #3