Re-instate BBOXX/Slate css after upstream merge
The old version of slate that we forked only works with:
    - ruby2.2
    - ruby2.2-dev
    - ruby-bundler == 1.12.5.

It's possible to install ruby2.2 from source but it's difficult to get
a valid version of bundler.

If you install newer version of ruby or ruby-bundler then Slate will
fail to compile the markdown, throwing this error:

`undefined method 'try' for "text/html; charset=utf-8":String`

This is a known issue with Slate (middleman):

(https://github.com/slatedocs/slate/issues/634)

Which is fixed in upstream/master. I've therefore merged in the latest
updates to Slate from upstream/master, updated ruby, ruby-env and
ruby-bundler to the more recent versions (2.5) and then re-applied the
"BBOXX Styling" that we have on our slate.

[JIRA: PL-401]