Update index.html.md
I can't seem to find a list of all of the "model" id's and their corresponding "model_name"s. Is there an endpoint that i can use to collect that list?

Is it possible to evaluate a URL against a single model via the API?

Are scores produced by the models on some comparable scale, such that a score produced by any model shares significance with the same score produced via any other model?

Are the scores returned in the API bounded between 0 and 1?

I believe we need an API key in order to get access, could we get an API key for the 100 articles a day, free tier plan while we build out the implementation.

Is the cost of our API related to the submission of a new url, or fetching the scores? (If I fetch the score of an article multiple times, does that count as usage or only the first time we submit a url).

 This last question might be me misinterpreting the api, but the results say "model": 1, "model": 2, is there a map between these values and what the model is predicting - as in "model": 1 -> "Political Bias", "model": 2 -> "threats", etc?